mejorando codigo
