+++ api/src/services/category.js
@@ -43,12 +43,9 @@ const getCategories = async (req, res) => {
       return res.status(200).send(
         await Category.findAll({
           include: {
-            model: Services,
-            attributes: ["id", "name"],
-            through: {
-              attributes: [],
-            },
-          },
+            model: User,
+            as: "user"
+          }
         })
       );
     }
@@ -61,23 +58,14 @@ const getCategories = async (req, res) => {
 
 const postCategorie = async (req, res) => {
   try {
-    const { name } = req.body;
-    if (!name) return res.send("Debes ingresar un nombre para la categoria");
-    const allCat = await allCategories();
-    const filterCategorie = allCat.filter((c) => c.name === name);
-    if (filterCategorie[0]) {
-      return res.send("Ya existe la categoria");
-    } else {
-      const createCategorie = await Category.create({
+    const { name, user_id } = req.body;
+      Category.create({
         name,
-      });
-      // let service = await Services.findAll({
-      //   where: { name: req.body.service },
-      // });
-      // createCategorie.addServices(service);
+        user_id
+      })
+    
       res.status(201).send("Category created");
-    }
-  } catch (error) {
+    } catch (error) {
     res.status(404).send(error);
   }
 };
