+++ client/src/componentes/Settings/EditProfile.jsx
@@ -1,129 +1,27 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { useAuth } from "../../context/authContext";
 import { useDispatch, useSelector } from "react-redux";
-import { getUserEmail, registerUser, updateUser } from "../../redux/actions";
-import axios from "axios";
-import { CLODUNIARY_API } from "../../Secret/Secret";
-import NavBar from "../PrivateRoute/Navbar";
+import { getUserEmail } from "../../redux/actions";
+
 import { Link } from "react-router-dom";
-
-function validate(fire) {
-    let error = {}
-
-    if(fire.birthDate < 18) error.birthDate = 'Debes ser mayor de edad'
-    // else if(typeof fire.phone !== 'number') error.phone = 'El numero ingresado es invalido'
-
-    return error
-}
+import CompleteProfile from "./AuxEditProfile/CompleteProfile";
+import UpdateProfile from "./AuxEditProfile/UpdateProfile";
 
 export default function EditProfile() {
   const { user } = useAuth();
   //ESTADO PARA ACTUALIZAR
   const estado = useSelector((state) => state.filter);
-  const [error, setError] = useState('')
-  const [input, setInput] = useState({
-    img: "",
-    firstName: "",
-    lastName: "",
-    phone: "",
-  });
-  const [fire, setFire] = useState({
-    email: "",
-    img: "",
-    firstName: "",
-    lastName: "",
-    phone: "",
-    birthDate: "",
-  });
-  const [btn, setBtn] = useState(false);
+
   const dispatch = useDispatch();
-  
+
   //PARA TRAER LA DATA DESDE LA BASE DE DATOS
 
   useEffect(() => {
     dispatch(getUserEmail(user?.email));
   }, [dispatch, user?.email]);
 
-  //ESTOS HANDLES SON PARA EL CASO EN QUE LOGEA CON FACEBOOK O GOOGLE
-  const handleOnChange = (e) => {
-    e.preventDefault();
-    setFire({
-      ...fire,
-      [e.target.name]: e.target.value,
-    });
-    setError(validate({
-        ...fire,
-      [e.target.name]: e.target.value,
-    }))
-  };
-
-  const handleOnSubmit = (e) => {
-    e.preventDefault();
-    if (fire.img === "") fire.img = user?.photoURL;
-    if (fire.email === "") fire.email = user?.email;
-    dispatch(registerUser(fire));
-    window.location.reload(true);
-  };
-  // PARA MANEJAR LOS NUEVOS DATOS INGRESADOS
-  const handleChange = (e) => {
-    e.preventDefault();
-    setInput({
-      ...input,
-      [e.target.name]: e.target.value,
-    });
-  };
-
-  const handleImage = async (e) => {
-    e.preventDefault();
-    try {
-      const file = e.target.files[0];
-      const data = new FormData();
-      data.append("file", file);
-      data.append("upload_preset", "changApp");
-
-      const cloudinary = await axios.post(CLODUNIARY_API, data);
-
-      setInput({
-        ...input,
-        img: cloudinary.data.secure_url,
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // PARA GUARDAR LOS CAMBIOS
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (input.img === "") input.img = estado[0].img;
-    if (input.lastName === "") input.lastName = estado[0].lastName;
-    if (input.firstName === "") input.firstName = estado[0].firstName;
-    if (input.phone === "") input.phone = estado[0].phone;
-    dispatch(updateUser(user?.email, input));
-    alert("Cambios guardados con existo");
-    window.location.reload(true);
-  };
-
-  //PARA CONTROLAR QUE SI NO INGRESO NINGUN DATO NO PUEDA GUARDAR LOS CAMBIOS
-  useEffect(() => {
-    if (input.img || input.firstName || input.lastName || input.phone) {
-      setBtn(false);
-    } else {
-      setBtn(true);
-    }
-  }, [input]);
-
-  //PARA CONTROLAR QUE INGRESE TODOS LOS DATOS (SINO SE ROMPE DESDE LA BASE DE DATOS)
-  useEffect(() => {
-    if (fire.lastName && fire.firstName && fire.birthDate && fire.phone && !error.birthDate && !error.phone) {
-      setBtn(false);
-    } else {
-      setBtn(true);
-    }
-  }, [fire, error]);
   return (
     <div>
-      <NavBar />
       <h1>Modificar perfil</h1>
       {user?.email === null ? (
         <p>
@@ -133,93 +31,10 @@ export default function EditProfile() {
           <Link to="/register">Registrarse</Link>
         </p>
       ) : estado.length === 1 ? (
-        <form onSubmit={(e) => handleSubmit(e)}>
-          <div>
-            <label>Imagen:</label>
-            
-            <input type="file" onChange={handleImage} placeholder="asd" />
-            {/* <img src={estado[0].img} alt="No se encontro" placeholder="asd" /> */}
-          </div>
-          <div>
-            <label>Nombre de Usuario: </label>
-            
-            <input
-              type="text"
-              value={input.firstName}
-              // defaultValue={estado[0].firstName}
-              placeholder={estado[0].firstName}
-              name="firstName"
-              onChange={handleChange}
-            />
-          </div>
-          <div>
-            <label>Apellido del Usuario:</label>
-            
-            <input
-              type="text"
-              placeholder={estado[0].lastName}
-              value={input.lastName}
-              name="lastName"
-              onChange={handleChange}
-            />
-          </div>
-          <div>
-            <label>Numero de Contacto: </label>
-            
-            <input
-              type="tel"
-              placeholder={estado[0].phone}
-              value={input.phone}
-              name="phone"
-              onChange={handleChange}
-            />
-          </div>
-          <button disabled={btn}>Guardar Cambios</button>
-        </form>
+        <UpdateProfile />
       ) : (
         <div>
-          <h4>Completar tus datos antes de seguir</h4>
-          <form onSubmit={(e) => handleOnSubmit(e)}>
-            <div>
-              <label>Nombre de usuario: </label>
-              <input
-                type="text"
-                name="firstName"
-                value={fire.firstName}
-                onChange={handleOnChange}
-              />
-            </div>
-            <div>
-              <label>Apellido del usuario: </label>
-              <input
-                type="text"
-                name="lastName"
-                value={fire.lastName}
-                onChange={handleOnChange}
-              />
-            </div>
-            <div>
-              <label>Fecha de nacimiento: </label>
-              <input
-                type="number"
-                name="birthDate"
-                value={fire.birthDate}
-                onChange={handleOnChange}
-              />
-            </div>
-            <div>
-              <label>Numero de telefono: </label>
-              <input
-                type="tel"
-                name="phone"
-                value={fire.phone}
-                onChange={handleOnChange}
-              />
-            </div>
-            <button type="submit" disabled={btn}>
-              Cargar Datos
-            </button>
-          </form>
+          <CompleteProfile />
         </div>
       )}
     </div>
